feat(radial-bar): isolate polar axes components in their own folder to easily move them to a dedicated package
