+++ packages/radial-bar/src/polar-axes/index.ts
@@ -1,2 +1,3 @@
+export * from './PolarGrid'
 export * from './RadialAxis'
 export * from './types'
