--- packages/radial-bar/src/polar_grid/index.ts
@@ -1,1 +1,0 @@
-export * from './PolarGrid'
